refactor : 재사용 컴포넌트 추출 및 경로 변경
