+++ fe/src/Post/components/PostInfo.jsx
@@ -1,13 +1,15 @@
 import React from 'react';
+import PostDetails from '../../components/PostDetails';
+import Partition from '../../components/Partition';
 import '../css/PostInfo.css';
 
-import day from '../img/day.png';
-import featherEye from '../img/feather_eye.png';
-import heart from '../img/heart.png';
+import date from '../img/day.png';
+import looked from '../img/feather_eye.png';
+import liked from '../img/heart.png';
 
 const result = {
   author: '에코노베이션',
-  day: '2022.05.04',
+  date: '2022.05.04',
   looked: '21',
   liked: '21',
 };
@@ -19,20 +21,11 @@ const PostInfo = function () {
         <div className="writer-img" />
         <span>{result.author}</span>
       </div>
-      <div className="spec-info">
-        <img src={day} alt="day" />
-        <span>{result.day}</span>
-      </div>
-      <span className="info-divider">|</span>
-      <div className="spec-info">
-        <img src={featherEye} alt="looked" />
-        <span>{result.looked}</span>
-      </div>
-      <span className="info-divider">|</span>
-      <div className="spec-info">
-        <img src={heart} alt="liked" />
-        <span>{result.liked}</span>
-      </div>
+      <PostDetails src={date} alt="date" info={result.date} />
+      <Partition />
+      <PostDetails src={looked} alt="looked" info={result.looked} />
+      <Partition />
+      <PostDetails src={liked} alt="liked" info={result.liked} />
     </div>
   );
 };
