+++ fe/src/components/Partition.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../css/Partition.css';
+import './css/Partition.css';
 
 const Partition = function () {
   return <span className="partition">|</span>;
