+++ fe/src/components/css/Partition.css
